+++ src/components/ListView/ListBox/ListBox.tsx
@@ -1,15 +1,15 @@
 import React, { useState, useEffect, useContext, useCallback } from 'react'
-import List from '@components/ListView/List/List'
 import { useDispatch, useSelector } from 'react-redux'
-import type { RootState } from '@stores/store'
-import { ListWrapper, SortBtns, ResultBox } from './ListBox.styles'
-import { ConvType } from '@stores/conv/convType'
-import { LIST_SORT_ITEMS } from '@utils/constants'
-import { distanceSort, reviewSort, starSort } from '@stores/conv/convSlice'
-import { saveSortType } from '@stores/sort/sortSlice'
-import KakaoService from '@services/kakaoService'
+import List from '@components/ListView/List/List'
+import LoadingWithLogo from '@components/styles/LoadingWithLogo'
 import { MapContext } from '@context/MapContext'
-import LoadingWithLogo from '@styles/LoadingWithLogo'
+import KakaoService from '@services/kakaoService'
+import { distanceSort, reviewSort, starSort } from '@stores/conv/convSlice'
+import { ConvType } from '@stores/conv/convType'
+import { saveSortType } from '@stores/sort/sortSlice'
+import type { RootState } from '@stores/store'
+import { LIST_SORT_ITEMS } from '@utils/constants'
+import { ListWrapper, SortBtns, ResultBox } from './ListBox.styles'
 
 const ListBox = () => {
   const sortedConv = useSelector((state: RootState) => state.conv.sortedStores)
