+++ src/services/markerImg.ts
@@ -1,13 +1,13 @@
-import GsMarker from '../assets/brandMaker/gsMaker.png'
 import cuMarker from '../assets/brandMaker/cuMaker.png'
-import miniMarker from '../assets/brandMaker/ministopMaker.png'
-import sevenMarker from '../assets/brandMaker/sevenMaker.png'
 import emartMarker from '../assets/brandMaker/emartMaker.png'
 import funMarker from '../assets/brandMaker/funlogMaker.png'
+import GsMarker from '../assets/brandMaker/gsMaker.png'
+import miniMarker from '../assets/brandMaker/ministopMaker.png'
 import myMarker from '../assets/brandMaker/myMarker.svg'
-import gsImg from '../assets/convImg/gs.png'
+import sevenMarker from '../assets/brandMaker/sevenMaker.png'
 import cuImg from '../assets/convImg/cu.png'
 import emartImg from '../assets/convImg/emart.png'
+import gsImg from '../assets/convImg/gs.png'
 import ministopImg from '../assets/convImg/ministop.png'
 import sevenImg from '../assets/convImg/seven.png'
