--- src/styles/Spinner.tsx
@@ -1,41 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-
-export const SpinnerOverlay = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`
-
-export const SpinnerContainer = styled.div`
-  display: inline-block;
-  width: 40px;
-  height: 40px;
-  border: 3px solid #fff;
-  border-radius: 50%;
-  border-top-color: #7d53d6;
-  margin-right: 30px;
-  animation: spin 1s ease-in-out infinite;
-  -webkit-animation: spin 1s ease-in-out infinite;
-  @keyframes spin {
-    to {
-      -webkit-transform: rotate(360deg);
-    }
-  }
-  @-webkit-keyframes spin {
-    to {
-      -webkit-transform: rotate(360deg);
-    }
-  }
-`
-
-const Spinner = () => {
-  return (
-    <SpinnerOverlay>
-      <SpinnerContainer className="spinner" />
-    </SpinnerOverlay>
-  )
-}
-
-export default Spinner
